refactor: move services interfaces into their respective directories
