+++ src/Commands/Meta/AbstractMetaSyncCommand.php
@@ -7,8 +7,8 @@ namespace AspirePress\AspireSync\Commands\Meta;
 use AspirePress\AspireSync\Commands\AbstractBaseCommand;
 use AspirePress\AspireSync\Integrations\Wordpress\WordpressApiConnector;
 use AspirePress\AspireSync\Resource;
-use AspirePress\AspireSync\Services\Interfaces\ListServiceInterface;
-use AspirePress\AspireSync\Services\Interfaces\MetadataServiceInterface;
+use AspirePress\AspireSync\Services\List\ListServiceInterface;
+use AspirePress\AspireSync\Services\Metadata\MetadataServiceInterface;
 use AspirePress\AspireSync\Utilities\StringUtil;
 use Exception;
 use Generator;
