+++ src/Services/Download/ThemeDownloadService.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AspirePress\AspireSync\Services\Download;
 
 use AspirePress\AspireSync\Services\Download\AbstractDownloadService;
-use AspirePress\AspireSync\Services\Interfaces\DownloadServiceInterface;
+use AspirePress\AspireSync\Services\Download\DownloadServiceInterface;
 use AspirePress\AspireSync\Services\Metadata\ThemeMetadataService;
 use GuzzleHttp\Client as GuzzleClient;
 use League\Flysystem\Filesystem;
