+++ src/Services/List/AbstractListService.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace AspirePress\AspireSync\Services\List;
 
-use AspirePress\AspireSync\Services\Interfaces\ListServiceInterface;
-use AspirePress\AspireSync\Services\Interfaces\MetadataServiceInterface;
+use AspirePress\AspireSync\Services\List\ListServiceInterface;
+use AspirePress\AspireSync\Services\Metadata\MetadataServiceInterface;
 use AspirePress\AspireSync\Services\Interfaces\RevisionMetadataServiceInterface;
 use AspirePress\AspireSync\Services\Interfaces\SubversionServiceInterface;
